requirements.txt
olefile==0.44
Pillow>=6.2.0
SQLAlchemy>=1.4.40
beautifulsoup4>=4.5.3
colorgram.py>=1.1.0
colorama>=0.3.9
requests>=2.25.1